Refactoring.
There is no automatic disconnection when the 'receiver' is destroyed because
it's a functor with no QObject. However, since 5.2 there is an overload which
adds a "context object". When that object is destroyed, the connection is
broken (the context is also used for the thread affinity: the lambda will be
called in the thread of the event loop of the object used as context).